+++ src/cluster_resources.rs
@@ -10,7 +10,10 @@ use crate::{
         },
     },
     error::{Error, OperatorResult},
-    labels::{APP_INSTANCE_LABEL, APP_MANAGED_BY_LABEL, APP_NAME_LABEL},
+    kvp::{
+        consts::{K8S_APP_INSTANCE_KEY, K8S_APP_MANAGED_BY_KEY, K8S_APP_NAME_KEY},
+        Label, LabelError, Labels,
+    },
     utils::format_full_controller_name,
 };
 
@@ -342,14 +345,25 @@ impl ClusterResource for DaemonSet {
 pub struct ClusterResources {
     /// The namespace of the cluster
     namespace: String,
+
     /// The name of the cluster
     app_instance: String,
+
     /// The name of the application
     app_name: String,
-    /// The manager of the cluster resources, e.g. the controller
+
+    // TODO (Techassi): Add doc comments
+    operator_name: String,
+
+    // TODO (Techassi): Add doc comments
+    controller_name: String,
+
+    // TODO (Techassi): Add doc comment
     manager: String,
+
     /// The unique IDs of the cluster resources
     resource_ids: HashSet<String>,
+
     /// Strategy to manage how cluster resources are applied. Resources could be patched, merged
     /// or not applied at all depending on the strategy.
     apply_strategy: ClusterResourceApplyStrategy,
@@ -395,6 +409,8 @@ impl ClusterResources {
             namespace,
             app_instance,
             app_name: app_name.into(),
+            operator_name: operator_name.into(),
+            controller_name: controller_name.into(),
             manager: format_full_controller_name(operator_name, controller_name),
             resource_ids: Default::default(),
             apply_strategy,
@@ -403,17 +419,15 @@ impl ClusterResources {
 
     /// Return required labels for cluster resources to be uniquely identified for clean up.
     // TODO: This is a (quick-fix) helper method but should be replaced by better label handling
-    pub fn get_required_labels(&self) -> BTreeMap<String, String> {
-        vec![
-            (
-                APP_INSTANCE_LABEL.to_string(),
-                self.app_instance.to_string(),
-            ),
-            (APP_MANAGED_BY_LABEL.to_string(), self.manager.to_string()),
-            (APP_NAME_LABEL.to_string(), self.app_name.to_string()),
-        ]
-        .into_iter()
-        .collect()
+    pub fn get_required_labels(&self) -> Result<Labels, LabelError> {
+        let mut labels = Labels::common(&self.app_name, &self.app_instance)?;
+
+        labels.insert(Label::managed_by(
+            &self.operator_name,
+            &self.controller_name,
+        )?);
+
+        Ok(labels)
     }
 
     /// Adds a resource to the cluster resources.
@@ -442,7 +456,11 @@ impl ClusterResources {
     ) -> OperatorResult<T> {
         Self::check_labels(
             resource.labels(),
-            &[APP_INSTANCE_LABEL, APP_MANAGED_BY_LABEL, APP_NAME_LABEL],
+            &[
+                K8S_APP_INSTANCE_KEY,
+                K8S_APP_MANAGED_BY_KEY,
+                K8S_APP_NAME_KEY,
+            ],
             &[&self.app_instance, &self.manager, &self.app_name],
         )?;
 
@@ -664,17 +682,17 @@ impl ClusterResources {
         let label_selector = LabelSelector {
             match_expressions: Some(vec![
                 LabelSelectorRequirement {
-                    key: APP_INSTANCE_LABEL.into(),
+                    key: K8S_APP_INSTANCE_KEY.into(),
                     operator: "In".into(),
                     values: Some(vec![self.app_instance.to_owned()]),
                 },
                 LabelSelectorRequirement {
-                    key: APP_NAME_LABEL.into(),
+                    key: K8S_APP_NAME_KEY.into(),
                     operator: "In".into(),
                     values: Some(vec![self.app_name.to_owned()]),
                 },
                 LabelSelectorRequirement {
-                    key: APP_MANAGED_BY_LABEL.into(),
+                    key: K8S_APP_MANAGED_BY_KEY.into(),
                     operator: "In".into(),
                     values: Some(vec![self.manager.to_owned()]),
                 },